fix(exports): interpret PN tags as union between object and string
the JSON standard represents PeopleNames as an object, rather than a delineated string.  this caused SRs derived
from metadata to have [object Object] for the patient name, notable in Google's Healthcare dicom store, which
seems to pick a random patient name when searching for series. much of the change is attempting to communicate
this union somewhat opaquely, and ensuring objects with additional properties aren't destroyed when being
naturalized. tested with OHIF master branch.